Consolidate Classification model
The Classification model was solely inherited by TopicalEvent. Therefore
moving the content of Classification into TopicalEvent.
