+++ app/assets/stylesheets/admin/_base.scss
@@ -30,7 +30,7 @@
 @import "fact-check-requests";
 @import "govspeak-help";
 @import "organisation";
-@import "classifications";
+@import "topical-events";
 @import "person";
 @import "sidebar";
 @import "lists";
