+++ app/assets/stylesheets/admin/_topical-events.scss
