+++ app/controllers/admin/classification_featurings_controller.rb
@@ -5,13 +5,13 @@ class Admin::ClassificationFeaturingsController < Admin::BaseController
     filter_params = params.slice(:page, :type, :author, :organisation, :title)
                           .permit!
                           .to_h
-                          .merge(state: "published", classification: @classification.to_param)
+                          .merge(state: "published", classification: @topical_event.to_param)
 
     @filter = Admin::EditionFilter.new(Edition, current_user, filter_params)
     @tagged_editions = editions_to_show
 
-    @classification_featurings = @classification.classification_featurings
-    @featurable_offsite_links = @classification.offsite_links
+    @classification_featurings = @topical_event.classification_featurings
+    @featurable_offsite_links = @topical_event.offsite_links
 
     if request.xhr?
       render partial: "admin/classification_featurings/featured_documents"
@@ -23,19 +23,19 @@ class Admin::ClassificationFeaturingsController < Admin::BaseController
   def new
     featured_edition = Edition.find(params[:edition_id]) if params[:edition_id].present?
     featured_offsite_link = OffsiteLink.find(params[:offsite_link_id]) if params[:offsite_link_id].present?
-    @classification_featuring = @classification.classification_featurings.build(edition: featured_edition, offsite_link: featured_offsite_link)
+    @classification_featuring = @topical_event.classification_featurings.build(edition: featured_edition, offsite_link: featured_offsite_link)
     @classification_featuring.build_image
   end
 
   def create
-    @classification_featuring = @classification.feature(classification_featuring_params)
+    @classification_featuring = @topical_event.feature(classification_featuring_params)
     if @classification_featuring.valid?
       flash[:notice] = if featuring_a_document?
-                         "#{@classification_featuring.edition.title} has been featured on #{@classification.name}"
+                         "#{@classification_featuring.edition.title} has been featured on #{@topical_event.name}"
                        else
-                         "#{@classification_featuring.offsite_link.title} has been featured on #{@classification.name}"
+                         "#{@classification_featuring.offsite_link.title} has been featured on #{@topical_event.name}"
                        end
-      redirect_to polymorphic_path([:admin, @classification, :classification_featurings])
+      redirect_to polymorphic_path([:admin, @topical_event, :classification_featurings])
     else
       render :new
     end
@@ -43,24 +43,24 @@ class Admin::ClassificationFeaturingsController < Admin::BaseController
 
   def order
     params[:ordering].each do |classification_featuring_id, ordering|
-      @classification.classification_featurings.find(classification_featuring_id).update_column(:ordering, ordering)
+      @topical_event.classification_featurings.find(classification_featuring_id).update_column(:ordering, ordering)
     end
-    redirect_to polymorphic_path([:admin, @classification, :classification_featurings]), notice: "Featured items re-ordered"
+    redirect_to polymorphic_path([:admin, @topical_event, :classification_featurings]), notice: "Featured items re-ordered"
   end
 
   def destroy
-    @classification_featuring = @classification.classification_featurings.find(params[:id])
+    @classification_featuring = @topical_event.classification_featurings.find(params[:id])
 
     if featuring_a_document?
       edition = @classification_featuring.edition
       @classification_featuring.destroy!
-      flash[:notice] = "#{edition.title} has been unfeatured from #{@classification.name}"
+      flash[:notice] = "#{edition.title} has been unfeatured from #{@topical_event.name}"
     else
       offsite_link = @classification_featuring.offsite_link
       @classification_featuring.destroy!
-      flash[:notice] = "#{offsite_link.title} has been unfeatured from #{@classification.name}"
+      flash[:notice] = "#{offsite_link.title} has been unfeatured from #{@topical_event.name}"
     end
-    redirect_to polymorphic_path([:admin, @classification, :classification_featurings])
+    redirect_to polymorphic_path([:admin, @topical_event, :classification_featurings])
   end
 
   helper_method :featuring_a_document?
@@ -71,14 +71,14 @@ class Admin::ClassificationFeaturingsController < Admin::BaseController
 private
 
   def load_classification
-    @classification = Classification.find(params[:topical_event_id] || params[:topic_id])
+    @topical_event = Classification.find(params[:topical_event_id] || params[:topic_id])
   end
 
   def editions_to_show
     if filter_values_set?
       @filter.editions
     else
-      @classification.editions.published
+      @topical_event.editions.published
                               .with_translations
                               .order("editions.created_at DESC")
                               .page(params[:page])
