+++ app/models/organisation.rb
@@ -79,7 +79,7 @@ class Organisation < ApplicationRecord
   has_many :organisation_classifications,
            -> { order("organisation_classifications.ordering") },
            dependent: :destroy
-  has_many :classifications, through: :organisation_classifications
+  has_many :topical_events, through: :organisation_classifications
 
   has_many :topical_events,
            -> { order("organisation_classifications.ordering") },
@@ -131,7 +131,7 @@ class Organisation < ApplicationRecord
   accepts_nested_attributes_for :default_news_image, reject_if: :all_blank
   accepts_nested_attributes_for :organisation_roles
   accepts_nested_attributes_for :edition_organisations
-  accepts_nested_attributes_for :organisation_classifications, reject_if: ->(attributes) { attributes["classification_id"].blank? }, allow_destroy: true
+  accepts_nested_attributes_for :organisation_classifications, reject_if: ->(attributes) { attributes["topical_event_id"].blank? }, allow_destroy: true
   accepts_nested_attributes_for :offsite_links
 
   validates :slug, presence: true, uniqueness: { case_sensitive: false }
