+++ app/views/admin/classification_featurings/index.html.erb
@@ -1,4 +1,4 @@
-<% page_title "Featured items on " + @classification.name %>
+<% page_title "Featured items on " + @topical_event.name %>
 
 <% content_for :current_tab do %>
   <h2>Currently featured</h2>
@@ -10,7 +10,7 @@
       To change the order of your featured documents, drag them up or down and then click "Save ordering".
     </p>
 
-    <%= form_tag [:order, :admin, @classification, ClassificationFeaturing], method: :put do %>
+    <%= form_tag [:order, :admin, @topical_event, ClassificationFeaturing], method: :put do %>
       <fieldset class="sortable">
         <%= render @classification_featurings %>
       </fieldset>
@@ -25,11 +25,11 @@
   <% end %>
 
   <hr />
-  <h2 class="feature-title add-bottom-margin">Feature new documents tagged with ‘<%= @classification.name %>’</h2>
+  <h2 class="feature-title add-bottom-margin">Feature new documents tagged with ‘<%= @topical_event.name %>’</h2>
 
   <div class="row">
     <div class="col-md-3">
-      <%= render partial: 'admin/editions/filter_options', locals: { filter_by: [:title, :type, :author, :organisation], filter_action: admin_topic_classification_featurings_url(@classification) } %>
+      <%= render partial: 'admin/editions/filter_options', locals: { filter_by: [:title, :type, :author, :organisation], filter_action: admin_topic_classification_featurings_url(@topical_event) } %>
     </div>
     <div class="col-md-9" id="search_results">
       <%= render partial: 'admin/classification_featurings/featured_documents' %>
@@ -42,7 +42,7 @@
   <% else %>
     <p class="no-content no-content-bordered">There are currently no non-GOV.UK government links.</p>
   <% end %>
-  <%= link_to "Create a non-GOV.UK government link", polymorphic_path([:new, :admin, @classification, :offsite_link]), class: "btn btn-default" %>
+  <%= link_to "Create a non-GOV.UK government link", polymorphic_path([:new, :admin, @topical_event, :offsite_link]), class: "btn btn-default" %>
 <% end %>
 
-<%= render partial: "admin/classifications/tab_wrapper", locals: { model: @classification } %>
+<%= render partial: "admin/topical_events/tab_wrapper", locals: { model: @topical_event } %>
