+++ app/views/admin/classification_featurings/new.html.erb
@@ -2,15 +2,15 @@
 
 <h2>
   <% if featuring_a_document? %>
-    Feature ‘<%= @classification_featuring.edition.title %>’ within ‘<%= @classification.name %>’
+    Feature ‘<%= @classification_featuring.edition.title %>’ within ‘<%= @topical_event.name %>’
   <% else %>
-    Feature an offsite document within ‘<%= @classification.name %>’
+    Feature an offsite document within ‘<%= @topical_event.name %>’
   <% end %>
 </h2>
 
 <div class="row">
   <div class="col-md-8">
-    <%= form_for [:admin, @classification, @classification_featuring], builder: Whitehall::FormBuilder, html: {class: 'well'} do |form| %>
+    <%= form_for [:admin, @topical_event, @classification_featuring], builder: Whitehall::FormBuilder, html: {class: 'well'} do |form| %>
       <%= form.errors %>
 
       <%= form.fields_for :image do |image_fields| %>
@@ -24,7 +24,7 @@
         <%= form.hidden_field :offsite_link_id %>
       <% end %>
 
-      <%= form.save_or_cancel cancel: polymorphic_path([:admin, @classification, :classification_featurings]) %>
+      <%= form.save_or_cancel cancel: polymorphic_path([:admin, @topical_event, :classification_featurings]) %>
     <% end %>
   </div>
 </div>
