--- app/views/admin/classifications/show.html.erb
@@ -1,51 +1,0 @@
-<% page_title @classification.name %>
-
-<% content_for :current_tab do %>
-  <table class="<%= model_name %> table table-bordered add-top-margin">
-    <tr>
-      <th>Summary</th>
-      <td class="summary"><%= @classification.summary %></td>
-    </tr>
-
-    <tr>
-      <th>Description</th>
-      <td class="description"><%= @classification.description %></td>
-    </tr>
-
-    <tr>
-      <th>Details</th>
-      <td class="details"><%= classification_contents_breakdown(@classification) %></td>
-    </tr>
-
-    <% if @classification.is_a?(TopicalEvent) %>
-      <tr>
-        <th>Duration</th>
-        <td class="duration">
-          <% if @classification.start_date && @classification.end_date %>
-            <%= @classification.start_date %> to <%= @classification.end_date %>
-          <% end %>
-        </td>
-      </tr>
-    <% end %>
-
-    <% if @classification.respond_to?(:featured_links) %>
-      <tr>
-        <th>Featured links</th>
-        <td class="featured-links">
-          <% if @classification.featured_links.any? %>
-            <%= render 'shared/featured_links', links: @classification.featured_links %>
-          <% else %>
-            <em>None</em>
-          <% end %>
-        </td>
-      </tr>
-    <% end %>
-  </table>
-
-
-  <div class="form-actions">
-    <%= link_to "Edit", [:edit, :admin, @classification], class: "btn btn-lg btn-primary"%>
-  </div>
-<% end %>
-
-<%= render partial: 'tab_wrapper', locals: { model: @classification } %>
