+++ app/views/admin/organisations/_form.html.erb
@@ -82,7 +82,7 @@
       <p>
         <%= label_tag "organisation_topical_event_ids_#{oc.ordering}" do %>
           Topical Event <%= oc.ordering + 1 %>
-          <%= select_tag "organisation[organisation_classifications_attributes][][classification_id]", options_from_collection_for_select(TopicalEvent.all, 'id', 'name', oc.classification_id), include_blank: true, multiple: false, class: 'chzn-select form-control', data: { placeholder: "Choose topical events…"}, id: "organisation_topical_event_ids_#{oc.ordering}" %>
+          <%= select_tag "organisation[organisation_classifications_attributes][][topical_event_id]", options_from_collection_for_select(TopicalEvent.all, 'id', 'name', oc.topical_event_id), include_blank: true, multiple: false, class: 'chzn-select form-control', data: { placeholder: "Choose topical events…"}, id: "organisation_topical_event_ids_#{oc.ordering}" %>
           <%= hidden_field_tag "organisation[organisation_classifications_attributes][][ordering]", oc.ordering %>
           <%= hidden_field_tag "organisation[organisation_classifications_attributes][][id]", oc.id %>
         <% end %>
