+++ app/views/admin/topical_event_about_pages/show.html.erb
@@ -34,4 +34,4 @@
   <% end %>
 <% end %>
 
-<%= render partial: "admin/classifications/tab_wrapper", locals: { model: @topical_event } %>
+<%= render partial: "admin/topical_events/tab_wrapper", locals: { model: @topical_event } %>
