+++ app/views/admin/topical_events/_custom_form_fields.html.erb
@@ -1,31 +1,31 @@
 <fieldset class="named logo">
   <legend>Logo</legend>
   <div class="form-group">
-    <%= image_tag classification_form.object.logo_url if classification_form.object.logo_url %>
-    <%= classification_form.text_field :logo_alt_text, label_text: 'Alt text' %>
-    <%= classification_form.upload :logo, label_text: 'File', allow_removal: classification_form.object.logo_url.present? %>
+    <%= image_tag topical_event_form.object.logo_url if topical_event_form.object.logo_url %>
+    <%= topical_event_form.text_field :logo_alt_text, label_text: 'Alt text' %>
+    <%= topical_event_form.upload :logo, label_text: 'File', allow_removal: topical_event_form.object.logo_url.present? %>
   </div>
 </fieldset>
 
 <fieldset class="named duration">
   <legend>Duration</legend>
   <div class="form-group">
-    <%= classification_form.label :start_date, "Start Date" %>
+    <%= topical_event_form.label :start_date, "Start Date" %>
     <div class="form-inline">
-      <%= classification_form.date_select :start_date, { include_blank: true, start_year: 2013 }, { :class => "date form-control" } %>
+      <%= topical_event_form.date_select :start_date, { include_blank: true, start_year: 2013 }, { :class => "date form-control" } %>
     </div>
   </div>
   <div class="form-group">
-    <%= classification_form.label :end_date, "End Date" %>
+    <%= topical_event_form.label :end_date, "End Date" %>
     <div class="form-inline">
-      <%= classification_form.date_select :end_date, { include_blank: true, start_year: 2013 }, { :class => "date form-control" } %>
+      <%= topical_event_form.date_select :end_date, { include_blank: true, start_year: 2013 }, { :class => "date form-control" } %>
     </div>
   </div>
 </fieldset>
 
 <fieldset class="named social_media_accounts">
   <legend>Social media accounts</legend>
-  <%= classification_form.fields_for :social_media_accounts do |account_form| %>
+  <%= topical_event_form.fields_for :social_media_accounts do |account_form| %>
     <div class="form-group">
       <%= account_form.label :social_media_service_id, 'Service' %>
       <%= account_form.select :social_media_service_id, options_from_collection_for_select(SocialMediaService.all, :id, :name, account_form.object.social_media_service_id), {include_blank: true}, class: 'chzn-select form-control', data: { placeholder: "Choose the social media service…" } %>
