+++ app/views/admin/topical_events/_form.html.erb
@@ -3,25 +3,25 @@
   show_consult_gds_warning ||= false
 %>
 
-<%= form_for [:admin, classification] do |classification_form| %>
-  <%= classification_form.errors %>
+<%= form_for [:admin, topical_event] do |topical_event_form| %>
+  <%= topical_event_form.errors %>
 
   <fieldset>
-    <%= classification_form.text_field :name %>
-    <%= classification_form.text_area :summary, rows: 3, required: true %>
-    <%= classification_form.text_area :description, rows: 20, class: 'previewable', data: {
+    <%= topical_event_form.text_field :name %>
+    <%= topical_event_form.text_area :summary, rows: 3, required: true %>
+    <%= topical_event_form.text_area :description, rows: 20, class: 'previewable', data: {
       module: "paste-html-to-govspeak"
     } %>
   </fieldset>
 
-  <%= render partial: 'custom_form_fields', locals: {classification_form: classification_form} %>
+  <%= render partial: 'custom_form_fields', locals: {topical_event_form: topical_event_form} %>
 
   <div class="row">
     <fieldset id="organisations" class="named col-md-6">
       <legend>Organisations</legend>
 
       <div id="organisation_sortable">
-        <%= classification_form.fields_for :organisation_classifications, classification.organisation_classifications.where(lead: false) do |organisation_form| %>
+        <%= topical_event_form.fields_for :organisation_classifications, topical_event.organisation_classifications.where(lead: false) do |organisation_form| %>
           <% organisation = organisation_form.object.organisation %>
           <div id="<%= dom_id(organisation) %>" class="well remove-bottom-padding">
             <%= organisation_form.text_field :lead_ordering, label_text: link_to(organisation.name, admin_organisation_path(organisation)), class: "ordering" %>
@@ -35,7 +35,7 @@
       <legend>Lead organisations</legend>
 
       <div id="lead_organisation_sortable">
-        <%= classification_form.fields_for :organisation_classifications, classification.lead_organisation_classifications do |lead_organisation_form| %>
+        <%= topical_event_form.fields_for :organisation_classifications, topical_event.lead_organisation_classifications do |lead_organisation_form| %>
           <% lead_organisation = lead_organisation_form.object.organisation %>
           <div id="<%= dom_id(lead_organisation) %>" class="well">
             <%= lead_organisation_form.text_field :lead_ordering, label_text: link_to(lead_organisation.name, admin_organisation_path(lead_organisation)), class: "ordering" %>
@@ -53,5 +53,5 @@
       Do not create <%= human_friendly_model_name.downcase %>s without consulting GDS.
     <% end %>
   </p>
-  <%= classification_form.save_or_cancel cancel: [:admin, classification] %>
+  <%= topical_event_form.save_or_cancel cancel: [:admin, topical_event] %>
 <% end %>
