+++ app/views/admin/topical_events/_tab_wrapper.html.erb
