+++ app/views/admin/topical_events/_topical_event.html.erb
@@ -4,7 +4,7 @@
   </td>
   <td class="summary"><%= truncate(topical_event.summary, length: 130, separator: ' ') %></td>
   <td class="description"><%= truncate(topical_event.description, length: 130, separator: ' ') %></td>
-  <td class="details"><%= classification_contents_breakdown(topical_event) %></td>
+  <td class="details"><%= topical_event_contents_breakdown(topical_event) %></td>
   <td class="duration">
     <% if topical_event.start_date && topical_event.end_date %>
       <%= topical_event.start_date %> to <%= topical_event.end_date %>
