+++ app/views/admin/topical_events/index.html.erb
@@ -16,6 +16,6 @@
   </thead>
 
   <tbody>
-    <%= render @classifications %>
+    <%= render @topical_events %>
   </tbody>
 </table>
