+++ app/views/admin/topical_events/new.html.erb
@@ -3,6 +3,6 @@
   <div class="col-md-8">
     <h1>New <%= human_friendly_model_name.downcase %></h1>
     <p class="warning">Do not create <%= human_friendly_model_name.downcase %>s without consulting GDS. New documents will be live immediately on selecting save.</p>
-    <%= render partial: "form", locals: {classification: @classification, show_consult_gds_warning: true} %>
+    <%= render partial: "form", locals: {topical_event: @topical_event, show_consult_gds_warning: true} %>
   </div>
 </div>
