--- test/factories/classifications.rb
@@ -1,6 +1,0 @@
-FactoryBot.define do
-  factory :classification do
-    sequence(:name) { |index| "classification-#{index}" }
-    description { "Classification description" }
-  end
-end
