+++ test/unit/admin/edition_filter_test.rb
@@ -190,11 +190,11 @@ class Admin::EditionFilterTest < ActiveSupport::TestCase
     assert_equal [newer_news_article], Admin::EditionFilter.new(Edition, @current_user, from_date: 2.days.ago.to_date.to_fs(:short)).editions
   end
 
-  test "can filter by classifications" do
+  test "can filter by topical_events" do
     topical_event = create(:topical_event)
     tagged_news   = create(:published_news_article, topical_events: [topical_event])
     _not_tagged   = create(:published_news_article)
-    filter        = Admin::EditionFilter.new(Edition, @current_user, classification: topical_event.to_param)
+    filter        = Admin::EditionFilter.new(Edition, @current_user, topical_event: topical_event.to_param)
 
     assert_equal [tagged_news], filter.editions
   end
