+++ test/unit/edition_test.rb
@@ -908,7 +908,7 @@ class EditionTest < ActiveSupport::TestCase
   test "republishes a linked Topical Event when the edition is changed" do
     edition = create(:edition, :draft)
     topical_event = create(:topical_event, :active)
-    create(:classification_featuring, classification: topical_event, edition: edition)
+    create(:classification_featuring, topical_event: topical_event, edition: edition)
 
     Whitehall::PublishingApi.expects(:republish_async).with(topical_event).once
     edition.update!(title: "some updated title")
